Add buildinfo.proto and its BUILD rule.
Summary:
This file adds a message to carry build tool information into the
details field of a CompilationUnit. This is mainly of interest for
Bazel, but may also apply to other tools that assign target names and
rule types.

Add generated code for buildinfo.proto.

Make proto generator work from anywhere.

It is now possible to run the generator script from anywhere under the
workspace root where a build is possible.

 - Add a shebang line and set the x bit on the script.
 - Find protobufs relative to the workspace, not the working directory.
 - Normalize quotation marks.
 - Wrap to 80 columns.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1047