Roll forward our minimum required LLVM/Clang version
Summary:
Changes in Clang related to Objective-C have made it difficult
to easily use the Kythe tools at our old Clang revision and at
top-of-tree Clang. This diff bumps the minimum revision up,
uses cmake to build LLVM instead of the (deprecated and removed)
autotools-based build system, and builds only the LLVM/Clang targets
that Kythe uses. We also limit the number of LLVM targets (e.g., for
inline assembler) we're interested in to a smaller set.

Reviewers: shahms, #core_team

Reviewed By: shahms, #core_team

Subscribers: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1051