Add Bazel build details to Go compilations.
Summary:
Thus far we've been recording the build target name in the VName
signature, but that really isn't a stable place to keep it. Now that we
have a designated home, use it.

Nothing depends on this yet; we should probably consider updating the
other Bazel extractors to do the same, though.

- Add a library for Kythe-aware protobuf Any values.

The protobuf ptypes library assumes all message types are defined in the
type.googleapis.com namespace, and uses a vendored copy of the generated
code. This package adds a wrapper that rewrites the URLs of Kythe
protobuf types, and emits the "correct" message type.

- Add additional protobuf package mappings.
- Update the go_protobuf import to commit 1f49d83d9aa00e6ce4fc8258c71cc7786aec968a.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1053