Update Bazel Go extractor for rule changes.
Summary:
Update vnames.json for Go rules.

- Map paths relative to bazel-out to have a separate root, with the
  config-specific subdirectory elided.

- Map the Go standard library location to a separate corpus.

Now that we have replaced our artisanal Go rules with the officially
supported Bazel ones, the mechanism for getting the build settings is
different. This change updates the extractor accordingly:

 - Drop the "Root" field of the extractor's Config struct.
 - Parse the Bazel response file for the real command-line arguments.
 - Filter out library inputs not used by the build settings.

Addresses T110 well enough to put it to rest for now.
We could still do better, but the result is manageable now.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Maniphest Tasks: T110

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1068