Add ability to skip certain tests on non-darwin machines.
Summary: Use configurable attribute to execute a different test script on darwin and Linux. This allows a single test to "skip" execution on non-darwin architecture. This is useful because we have some tests that only pass on darwin. The other option is to add a tag so they aren't executed as part of a group or during pre-submit. This solution allows us to keep the test in groups and pre-submit so it is still actually tested as long as it is possible to test it.

Test Plan: The property_decl_defn_xcode_only test previously failed on Linux. Now it passes trivially. I made property_decl_defn_xcode_only.m have a bad assertion and verified that it would fail on darwin, meaning the configurable attribute was causing the real test to run on darwin; it was not making use of the always-pass test script.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1096