Add category and extension support.
Summary:
Categories will be treated as a special kind of extends relationship. The
methods defined in a category will belong to the category object, not the base
class. Extensions will be treated in a similar way but since they do not have a
name and are not referred to in the impl block, they will be missing some
features (like the ability to connect the impl directly to the extension).

Also, update lint script to make it safe for files and targets with "+" in their name.

Test Plan: Added a test for a category and a test for an extension.

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1091