Add compatibility_alias support and other tests
Summary:
Add support for Compatibility alias (ObjCCompatibleAliasDecl). This is
similar to a typedef. The way clang handles it is slightly different
though because we don't see the alias in the AST, we just see the
aliased type, even if there are intermediate steps between the alias and
the root aliased type. See tests for examples and further explanation.

Add tests for the following no-op cases:
* encode expr
* subscript ref expr
* selector expr

Test Plan: New indexer tests.

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1099