Treat all builtin FunctionDecls as definitions
Summary:
A "builtin" node may have a declaration and a definition. Since we
create node IDs for builtin functions in a special way, we will get
an error if we record a declaration (incomplete) and a definition
(complete). To fix this, we treat all builtin functiondecls as definitions.

Test Plan: objc bazel extractor test will now exercise this case (but it only runs manually).

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1131