Split Nodes/Edges out of XRefService.
Summary:
Creates a new GraphService interface containing these methods.  For now
the existing implementations will satisfy both, but in general an
XRefService implementation will not be required to satisfy GraphService.

This commit also cleans up some of the preliminary design comments on
the XRefService definition and removes a related TODO.

Update usage sites for the XRefService.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1142