Make it more clear what a system header prefix is
Summary:
You can follow the uses of this value to get the whole story, but it is
used to determine if a header path is a system include. It does not
specify a path to look for system include headers.

For one location to look at to see why this is true, see
llvm/llvm/tools/clang/lib/Lex/HeaderSearch.cpp at about line 765 where
it is using SystemHeaderPrefixes to set if a directory is a system
include or not.
http://llvm.org/reports/coverage/tools/clang/lib/Lex/HeaderSearch.cpp.gcov.html
at line 760.

Test Plan: none, just text change

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1144