Pass env vars from spawn action info to extractor
Summary: Bazel sets env vars to execute the primary action. For extra actions, it passes them in as part of the proto we read. This means that we should set them if we are going to run any external commands so they will be run in the same environment as if they were run during the primary action.

Test Plan: Updated the test to add test cases for the new helper method.

Reviewers: shahms, zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: fromberger, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1140