Report source locations for references arising from implicit variable declarations
Summary:
Unfortunately D838 wasn't quite enough to correctly report a source
location for the references to 'begin', 'end' and the operators arising
from a range-for loop. An earlier change, D831, caused the indexer to stop
reporting source locations for references contained within an implicit
declaration. Although that may be a sensible thing to do in general, it
would seem to be inappropriate for references arising from range-for loops.

This patch reverts to the previous behavior if the reference was contained
within an implicit variable declaration, which we assume means that the
declaration arose from a range-for loop.

This unbreaks the //kythe/cs/testdata:for_range test in D807.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1172