Go indexer: Blame top-level entries on the package.
Summary:
Rather than blaming function calls and anonymous functions at file scope on the
file, blame them on the package initializer, which is now emitted as a dummy
function node. This node is created lazily on demand so that we won't have an
isolate in packages that don't need it.

Related
- Use 'package' as the signature for a Go package.
- Set up a dummy function for the package initializer.
- Update tests for blame attribution change.
- Add unit test for package associations.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1224