Go indexer: Add anchors & references for imports & imported packages.
Summary:
Imports are references (ref/imports) to their imported packages.

Uses of the package name on the LHS of a selector expr (pkg.Name) are plain
references (ref) to the imported package.  We do not currently record a
relationship between the ref anchor and the import site, as Kythe does not have
a separate node type for the import event itself.

Related:
- Add a unit test for import associations.
- Add ref/imports edge kind to the schema library.
- Add writeRef method to the emitter.

Cosmetic fixes:
- Remove a completed TODO comment.
- Rename parentFunc to stackFunc.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1237