Fix a panic in the gsFlag String implementation.
Summary:
The Go flag package instantiats a zero of the flag type for rendering purposes
(since 1.6 or 1.7, I think), and that leads to a nil indirection in the String
implementation. Check for the nil first to avoid this.

The Go team closed my https://github.com/golang/go/issues/16694 as WAI, though
I still argue this was a spurious violation of the compatibility guarantee.  So
we work around it.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1242