Fix flaky behavior in fyi
Summary:
Something suspicious is happening with fyi's handling of
builtin headers. I've changed the code to do something safe
(copy rodata strings instead of point to them in a way that
the API says should not cause them to be destroyed).

Filed T202 against myself to look at this in greater depth.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1246