Update Go proto libraries for latest proto3 support
Summary:
This change has to handle a new output filename for the any_proto in the proto.bzl rules.  If a proto file has a "go_package", it magically renames the output path so we must pipe it through and move the file to the expected location.  We could do something fancier to remove the need to specify the path, but this works for the one-off location we need it.

The latest Java/Go GRPC libraries also don't seem to work well together since the analyzer_driver test is now timing out while connecting to the Java analyzer.  We don't really rely on the GRPC support so I'm going to leave this as a TODO for now.

Reviewers: fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1249