Go indexer: Handle "wrapper" declarations for struct & interface.
Summary:
In a type declaration, "type A B" where B is an existing struct or interface
type generates the same type, but does not have the defining syntax.  Check
that we do before attempting to add explicit field or method edges.

Also: Set pipefail on the extractor and verifier actions.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1250