Go indexer: Load import data lazily to reduce file I/O.
Summary:
Prior to this commit, the resolver eagerly loaded every required input fully.
This is actually overkill, particularly in Bazel where each compile action is
sent the complete list of standard library packages.

This commit moves the loading step into the importer, and only loads those
packages explicitly needed by the compiler. This reduces the number of input
files that need to be fetched, saving time and memory.

Also: Fix a typo in a log message.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1255