Add an experimental option for aliasing templates together.
Summary:
This experiment takes a hacksaw to template instantiations.
Its goal is to see whether we can build reasonable databases
with templates turned on.

Note that there is no attempt at early-stage deduplication here
(which will definitely increase as a result of the smearing).
This change aims only at reducing the ultimate size of the graph.

Reviewers: fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1261