Permit more than one kindex/index pack job; add an option to squelch output.
Summary:
This diff is a step toward being able to write claiming tests
that don't require intermediate files (beyond .kindexes) or
standing up memcaches. The new silent option is meant to be used
to preload a claim database so that the output from subsequent
index files can be checked for the presence or absence of data.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1262