Update completeAnchors to remove use of "childof" edges.
Summary:
Because of the rules for http://kythe.io/docs/schema#anchors, it is not
necessary to read "childof" edges to connect an anchor to its enclosing file.
Simplify the completeAnchors function based on that assumption.

This is a step toward allowing us to drop explicit "childof" edges from each
anchor to its enclosing file.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1264