Prune the AST where possible.
Summary:
Dynamic claiming prunes too much of the AST because it
assumes that transcripts completely determine the shape of
(sub)trees. This isn't the case with template instantiations.
This diff (1) identifies which part of the AST is prunable by
marking which nodes contain template instantiations as subtrees
and (2) uses the existing claiming infrastructure to cut off
traversal at prunable nodes.

Reviewers: fromberger

Reviewed By: fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1269