Go extractor: Avoid double-rooting generated source paths.
Summary:
The Bazel rules for Go prefix generated source files with a label that the
compiler expects to find. Ensure we don't remove that label and thus confuse
the indexer.

Also: Handle packages with no source inputs gracefully in the indexer.

A package with no source files is usually a sign of a bug, typically in the
extractor. The indexer now reports an error rather than crashing, when this
occurs.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1272