Cish indexer should always emit c++ facts
Summary:
This indexer is used to index C, C++, and Objective-C. It is not
beneficial to use the exact language name in the facts that we output.

The possible advantages for using different names include a more precise
graph and more ability to specialize the post-processing pipeline.

The disadvantages of using different names are a need to create cross
language links in seemingly single-language code bases and possible
duplication of indexing.

The advantages sound like they might outweigh the disadvantages, but
the advantages are theoretical and the disadvantages would cost a lot of code
complexity.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1296