start a typescript test suite
Summary:
Add a program that runs the validator on a directory of tests.
This program is written in TypeScript so that it can reuse the
cached TypeScript standard library across tests.

Implement a tiny bit more of indexer to get the test to pass.
This indexer is still no good.  In particular, the signature for a
symbol used here is wrong, but it serves as a starting point.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1310