typescript: start on indexing functions
Summary:
Change the internal API around a bit so that there isn't a
function that accepts a plain string "signature", as that is
too easy to misuse (was accidentally passing an identifier
string in a place that wanted a signature).
Instead most functions should operate on VNames.

Implement the beginning of visitFunctionDeclaration, using
getSymbolName() (which also returns a VName) consistently.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1313