Add refs for obj-c method and receiver for class methods.
Summary:
We have a ref/call edge for obj-c methods, but we should also have a ref
edge for just the method name. Since the "method name" is not a
contiguous span in the file, we say the range for the ref/call is the
entire [...] span and the span for the ref is the selector. So in
[foo add:4 sub:29], the ref/call span is "[foo add:4 sub:29]" and the ref
span is "add:4 sub:29".

We were not getting refs for the receiver with class methods, so I
explicitly added a ref to the class at these call sites. So if we have
[Foo classMethod], we now have a ref from Foo to the class Foo.

make error more explicit

Test Plan: Added a few tests for these refs.

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1320