Added Objective-C multi-line test
Summary:
It is common for Objective-C methods to be split across multiple lines.
This situation is different than in other language because the actual
identifier for the method is split across multiple lines.

In Objective-C you might have:
-(int) foo:(int)a
       extra:(int)b;

Where you want to link foo:extra: to an anchor. A long term solution
would be to create separate links for each selector component.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1338