Log if class receiver is null
Summary:
While indexing some code, we were getting assertions errors that QT was
null. Add logging to figure out why this happens. The documentation at
https://clang.llvm.org/doxygen/classclang_1_1ObjCMessageExpr.html#a6457400eb37b74bf08efc178d6474e07
says that null is returned if the expression is not a class message, but
it doesn't promise that null is never returned if the expression is a
class message.

Reviewers: zarko

Reviewed By: zarko

Subscribers: zarko, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1341