Make file data provider interface more precise
Summary:
The file data provider interface was not precise on how the two
parameters would be used, so I wrote documentation that matches what the
implementations were already doing. They both could be null and it is up
to the implementation to handle that since each implementation uses the
parameters in different ways.

There may be a few functional changes here since we are validating input
and throwing early if it is null.

I also added a test for FileDataCache. I wanted to add tests for the
other file data providers, but they are hardwired to file system
implementations that are too much work to abstract in this diff.

Reviewers: craigbarber

Reviewed By: craigbarber

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1345