Stop emitting type ranges for builtin structural types.
Summary:
It's not useful in practice to emit `T*` as `[[T]*]`. Instead,
we should emit it as `[T]*`, where `T` refers to the type `T` (and not `T*`).
`T*` is still emitted as a type node and can be reached through
deeper graph queries. This argument extends to `T&`, `const T`, and so on.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: salguarnieri, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1347