Create type node for @interface if @implementation isn't present
Summary:
When we are creating a type node, the interface should be used if the
implementation isn't present. Only if there is no interface and no
implementation (there is only a forward @class declaration) should we
link to a nominal type node.

Add a test that makes sure we are linking to the interface node if the
implementation node isn't present.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1358