typescript: always build when running 'yarn test'
Summary:
You need to 'yarn build' before you run tests, and it's easy
to overlook.  Make the default behavior always build, and provide
'yarn runtest' for using while developing.

(Really this screams for a real build system but it's so simple
that hopefully this hacky thing will continue to work for now...)

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1359