typescript: begin handling type/value separation
Summary:
There are two namespaces of TypeScript: types and values.
A given symbol may be a type, it may be a value, and the two may even
be unrelated.  Handle this by giving the two separate vnames.

Add a test that uses the same name for a type and unrelated value, and
check that the various uses refer to the appropriate places.

(Note: there's a third namespace, namespaces.  To be handled later.)

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1360