Use first selector for source range of objc method decl and call ref
Summary:
Simplify the links we generate for objective c messages. By using the
range of the first selector instead of the entire range of the method
decl/selector, we are more similar to other languages where the method name is
completely independent of the parameters.

In the future, we should think about recording multiple ranges for objective c
methods (one for each part of the selector).

This change introduces some confusion. For example:

[foo bar: 10 baz: 20] will have the source range for bar (and only bar) ref the method decl
for bar:baz:.

[foo bar: 10 baz: 20 time: 40] will also have the source range for bar (and only bar) ref the
method decl for bar:baz:time:.

Test Plan: updated all tests to conform to the new style

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1363