Ensure that signatures involving Java wildcards are stable.
Summary:
Previously wildcards were disambiguated using a Java hashCode,
which is not stable across compiler invocations.

This change replaces that hashCode with a string composed of
the source file path (relative to the source root), and the index
of the wildcard in the order encountered in the source file.

This guarantees that the relevant vnames are stable across
multiple invocations, and even stable if a file is compiled
in different compilation units (e.g., compiling A.java and B.java
together, and later compiling each separately).

Reviewers: salguarnieri

Reviewed By: salguarnieri

Subscribers: salguarnieri, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1364