Record Objective-C Containers as classes
Summary:
In C++ we use NameEqClass to record if a decl is a class, macro, etc.
Now we also do this for Objective-C. In the future we might want to
provide different kinds for the different container subtypes (interface,
implementation, category, and protocol).

I also added a forward declare test to c++ that mirrors a test in
Objective-C to make sure both languages produced similar graphs. I
didn't mention that the two tests were related though because that isn't
important going forward because the two tests are not compared to each
other by our testing infrastructure.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1372