Add a Canonicalize method to the kcd.Unit type.
Summary:
When storing compilations, it is preferred to choose a canonical form to avoid
spurious differences due to irrelevant order constraints. To express this, add
a method the Unit can use to put its house in order. It is safe for this to be
a no-op; at worst it will lead to storing spuriously-redundant records.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1381