Specify an ordering rule for VNames.
Summary:
This affects the definition of Entry order, which is defined in terms of the
order of source/target fields, themselves vnames. When we first wrote this, we
had those fields as tickets; now that they're vnames it's underspecified.

Add vnameutil.Compare, a library to compare vnames according to this rule.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1383