serving.proto: add PagedCrossReferences.feature field
Summary:
The `feature` field lets the server know what kind of data is populated
in each PagedCrossReferences and whether it must compute some data for a
CrossReferencesReply dynamically.

This change shows how this may be used by adding the RELATED_NODES and
MARKED_SOURCE serving data features used to indicated whether each
PagedCrossReferences contains known RelatedNodes and a MarkedSource,
respectively.  If a feature is present in the serving data, the slow
dynamic paths usually used to compute the relevant data will be skipped
entirely.

Reviewers: zarko, fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1391