Add tests for ref edge from class "field"
Summary:
These tests are primarily for Objective-C to make sure a property
correctly references the type of that property. For example in
"@property Foo var", we want to make sure Foo refers to the decl/defn
for Foo.

If the class being referenced is a forward declaration, we must ref the
nominal node since we don't have anything better to refer to.

Reviewers: zarko

Reviewed By: zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1393