typescript: prepare for multi-file indexing
Summary:
Make the computation of signatures ignore the source file.  The full
vname of the node will include the source file path, so the signature
does not need to reference the path.

Change a few functions to work with arrays of files rather than single
files.

This has no effect on tests, but the observed vnames no longer have
a TODO in them.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: salguarnieri, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1394