typescript: fix test suite library caching
Summary:
The test suite caches the parse of the standard library across
each test.  It does so by relying on the output of
  compilerHost.getDefaultLibFileName
which is keyed off the "target" option.  When I changed the default
library in the options without changing the target, the test was
no longer caching the correct standard library version.

Fix this by using the "target" option to specify ES6 rather than
specifying the library by name.  This makes each test (after the
first) roughly twice as fast.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1398