typescript: don't generate invalid json on large outputs
Summary:
process.exit() doesn't ensure process.stdout has been flushed
before exiting, so if the indexer produced enough output it was
possible for it to exit before writing the full JSON output.

The fix is to set the exit code and let Node choose when to exit.

This fixes running the indexer on large files such as lib.es6.d.ts.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1405