typescript: use tsconfig.json for loading options/files
Summary:
In a production TypeScript build, compilation is driven by an input
tsconfig.json.  Change the indexer and test suite to both run from
one as well, and to index all files in the compilation by default.
This allows us to delete our local ts.CompilerOptions as well as
scanning for files.

It also changes the debug command "yarn browse" to browse the whole
project by default (both indexer and test are indexed).

Reviewers: #core_team, salguarnieri, zarko

Reviewed By: #core_team, salguarnieri, zarko

Subscribers: zarko, salguarnieri, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1410