typescript: only type-check the requested input files
Summary:
The indexer is passed a set of files to index; these files may
reference other files not in that list (e.g. the TypeScript standard
library).  Avoid type-checking these extra files as they're not
part of the indexing process.

This behavior is consistent with how we compile TypeScript code in
google3, and has the nice side effect of dramatically speeding up
the test suite because we no longer type-check the standard library
on each test.

Add a new file to the test suite that intentionally doesn't compile.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1439