Add max supported version of bazel
Summary: We have been broken by new releases of bazel a few times. The errors that occur in these situations are confusing and it is not obvious that the cause is a new Bazel release. This will make it clear what versions of Bazel we have tested our build with.

Reviewers: fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1440