Fix JDK (and other) cross-references
Summary:
This diff adds a flag to force consistent corpus
names for the JDK (should your build system require it).

It also fixes the way hashes are computed for
Java classes by dropping methods that are automatically
generated by javac. These methods are inconsistently
present depending on whether the class was loaded from
a .class file or built from .java source. For example,
java.lang.String#compareTo(java.lang.String) exists in
String.java (and must exist because String implements
Comparable<String>). After type erasure, java.lang.String
will implement Comparable, and therefore must implement
#compareTo(java.lang.Object). javac will synthesize a
[[ https://docs.oracle.com/javase/tutorial/java/generics/bridgeMethods.html | bridge method ]]
to satisfy this requirement. The bridge method appears in the
.class file (marked with the bytecode equivalent of the SYNTHETIC | BRIDGE flags in the diff),
but will not appear to the instance of javac compiling the .java file.
Just as we rely on dropping private members to get consistent hashes
between the .java and its compiled (and stripped) .class form (since
private members are visible in the .java context and not the .class),
so to must we drop SYNTHETIC/BRIDGE methods (as these
are visible in the .class but not in the .java). The implementation
here is an overapproximation (we look for SYNTHETIC | BRIDGE,
not SYNTHETIC & BRIDGE), but this is meant to guard against
cases where other methods are synthesized that we haven't yet
discovered.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: jrtom, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1442