Add a JSON proxy wrapper for proto-incapable indexers.
Summary:
Some languages have difficulty handling protocol buffers, or at least find it
much easier to work in JSON. To permit a driver to work with such languages,
this change adds a "proxy" library that converts the interaction into a simple
JSON protocol conducted over pipes.

This hasn't yet been integrated into the existing driver tool, but I'd like to
do that as a separate step. The typical mode of operation where we run the
indexer as a subprocess and pass work to it will continue to work in this new
regime, only the communication channel changes.

Reviewers: #core_team, zarko, evmar

Reviewed By: #core_team, zarko, evmar

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1443