Add a new "satisfies" edge kind to the Kythe schema.
Summary:
See T224. In Go, types may implicitly satisfy interfaces.  Unlike with explicit
extension (class A extends B {...}), the set of possible implicit satisfactions
can't be constructed without a global scan to discover what interfaces exist.

In practice, however, it is useful to record the relationship a concrete type T
has to "known" interfaces, particularly those occurring in compilations T
depends on.  Because this relationship is useful, but heuristic, let's give it
a separate name from "extends" (even though they are similar in spirit) so that
tools can tell whether the information in the graph is likely to be complete.

This change updates the schema and the Go indexer, and adds examples.

Reviewers: craigbarber, #core_team

Reviewed By: craigbarber, #core_team

Subscribers: craigbarber, #core_team

Maniphest Tasks: T224

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1445