Check for singleton EVars in the verifier.
Summary:
This diff adds a flag to enable singleton checking
(where the verifier will consider it an error if
an EVar is mentioned only once). It also allows nonce
EVars (_) to be arbitrary identifiers (_etc), provided
that they start with _. EVars that are inspected explicitly
(with ?) are also excluded from singleton checking.

Checking for singletons is not on by default, but once we
fix up the rest of the repo, we'll turn it on.

Reviewers: fromberger, #core_team

Reviewed By: fromberger, #core_team

Subscribers: pludemann, #core_team

Maniphest Tasks: T225

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1446