Ensure that the C++ indexer always has a job
Summary:
There is an invariant that the C++ indexer always
has an IndexJob while doing work. This invariant
was violated when the indexer was indexing top-level
comments. This diff restores the invariant by
turning that operation into a different kind of job.

This diff also removes a noisy log operation from
ProtoLibrarySupport.cc. It's not an error case; it's
a case where the plugin can't handle certain kinds
of input.

Reviewers: #core_team, fromberger

Reviewed By: #core_team, fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1452